Saturday Updates
created beginning of something to work on in greater detail on Monday. On Monday will work to copy and modify the ql_agent.py in another folder. Then I will work on allowing it to have either multiple runs on same environment or working with different environments. This is because the current agent in the file is very constricted.

Will then copy and paste that modified class into the main file.