final updates for 7/27
worked on a class to stop a lot of the clutter that was happening in the main function.

Am working on other things to study from this as well as making the output not just in memory and in the console, but finding a good way to keep it saved for later, if parts of this test want to be saved.